+++ lightstep/instrumentation/cputime/cputime.go
@@ -16,13 +16,8 @@ package cputime // import "github.com/lightstep/otel-launcher-go/lightstep/instr
 
 import (
 	"context"
-	"fmt"
-	"math"
-	"runtime"
-	"syscall"
 	"time"
 
-	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/metric"
 	"go.opentelemetry.io/otel/metric/global"
@@ -162,36 +157,4 @@ func (c *cputime) register() error {
 	}
 
 	return nil
-}
-
-// getProcessTimes calls ReadMemStats() for GCCPUFraction because as
-// of Go-1.19 there is no such runtime metric.  User and system sum to
-// 100% of CPU time; gc is an independent, comparable metric value.
-// These are correlated with uptime.
-func getProcessTimes() (userSeconds, systemSeconds, gcSeconds, uptimeSeconds float64) {
-	// Would really be better if runtime/metrics exposed this,
-	// making an expensive call for a single field that is not
-	// exposed via ReadMemStats().
-	var memStats runtime.MemStats
-	runtime.ReadMemStats(&memStats)
-
-	gomaxprocs := float64(runtime.GOMAXPROCS(0))
-
-	uptimeSeconds = time.Since(processStartTime).Seconds()
-	gcSeconds = memStats.GCCPUFraction * uptimeSeconds * gomaxprocs
-
-	var ru syscall.Rusage
-	if err := syscall.Getrusage(syscall.RUSAGE_SELF, &ru); err != nil {
-		userSeconds = math.NaN()
-		systemSeconds = math.NaN()
-		otel.Handle(fmt.Errorf("getrusage: %w", err))
-		return
-	}
-
-	utime := time.Duration(ru.Utime.Sec)*time.Second + time.Duration(ru.Utime.Usec)*time.Microsecond
-	stime := time.Duration(ru.Stime.Sec)*time.Second + time.Duration(ru.Stime.Usec)*time.Microsecond
-
-	userSeconds = utime.Seconds()
-	systemSeconds = stime.Seconds()
-	return
 }
