+++ lightstep/instrumentation/cputime/cputime_other.go
@@ -1,0 +1,59 @@
+// Copyright The OpenTelemetry Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+//go:build !(linux || darwin)
+
+package cputime // import "github.com/lightstep/otel-launcher-go/lightstep/instrumentation/cputime"
+
+import (
+	"fmt"
+	"math"
+	"runtime"
+	"syscall"
+	"time"
+
+	"go.opentelemetry.io/otel"
+)
+
+// getProcessTimes calls ReadMemStats() for GCCPUFraction because as
+// of Go-1.19 there is no such runtime metric.  User and system sum to
+// 100% of CPU time; gc is an independent, comparable metric value.
+// These are correlated with uptime.
+func getProcessTimes() (userSeconds, systemSeconds, gcSeconds, uptimeSeconds float64) {
+	// Would really be better if runtime/metrics exposed this,
+	// making an expensive call for a single field that is not
+	// exposed via ReadMemStats().
+	var memStats runtime.MemStats
+	runtime.ReadMemStats(&memStats)
+
+	gomaxprocs := float64(runtime.GOMAXPROCS(0))
+
+	uptimeSeconds = time.Since(processStartTime).Seconds()
+	gcSeconds = memStats.GCCPUFraction * uptimeSeconds * gomaxprocs
+
+	var ru syscall.Rusage
+	if err := syscall.Getrusage(syscall.RUSAGE_SELF, &ru); err != nil {
+		userSeconds = math.NaN()
+		systemSeconds = math.NaN()
+		otel.Handle(fmt.Errorf("getrusage: %w", err))
+		return
+	}
+
+	utime := time.Duration(ru.Utime.Sec)*time.Second + time.Duration(ru.Utime.Usec)*time.Microsecond
+	stime := time.Duration(ru.Stime.Sec)*time.Second + time.Duration(ru.Stime.Usec)*time.Microsecond
+
+	userSeconds = utime.Seconds()
+	systemSeconds = stime.Seconds()
+	return
+}
